backend/requirements.txt
django==3.0.5
djangorestframework==3.11.0
django-filter==2.2.0
django_extensions==2.2.9
django_cors_headers==3.2.1

psycopg2-binary==2.8.5
